+++ web/app/scripts/selected-visualization.js
@@ -1,0 +1,13 @@
+'use strict';
+
+angular.module('biggraph').directive('selectedVisualization', function() {
+  return {
+    restrict: 'E',
+    scope: {
+      side: '=',
+      visualization: '=',
+      title: '=',
+    },
+    templateUrl: 'selected-visualization.html',
+  };
+});